core requisite w3 act2 DONE
