+++ src/index.html
@@ -68,7 +68,7 @@
           </a>
         </li>
         <li class="product-card">
-          <a href="product_pages/northface-talus-4.html">
+          <a href="product_pages/index.html?product=985RF">
             <img
               src="images/tents/the-north-face-talus-tent-4-person-3-season-in-golden-oak-saffron-yellow~p~985rf_01~320.jpg"
               alt="Talus Tent - 4-Person, 3-Season" />
@@ -79,7 +79,7 @@
         </li>
         </li>
         <li class="product-card">
-          <a href="product_pages/northface-alpine-3.html">
+          <a href="product_pages/index.html?product=985PR">
             <img
               src="images/tents/the-north-face-alpine-guide-tent-3-person-4-season-in-canary-yellow-high-rise-grey~p~985pr_01~320.jpg"
               alt="Alpine Guide Tent - 3-Person, 4-Season" />
@@ -91,7 +91,7 @@
           </a>
         </li>
         <li class="product-card">
-          <a href="product_pages/cedar-ridge-rimrock-2.html">
+          <a href="product_pages/index.html?product=344YJ">
             <img src="images/tents/cedar-ridge-rimrock-tent-2-person-3-season-in-rust-clay~p~344yj_01~320.jpg"
               alt="Rimrock Tent - 2-Person, 3-Season" />
             <h3 class="card__brand">Cedar Ridge</h3>
